requirements.txt
baidu-aip>=2.2.2.0
behave>=1.2.6
certifi>=2018.1.18
chardet>=3.0.4
colorama>=0.3.9
cookies>=2.2.1
h2==2.6.2
hpack>=3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna>=2.6
lxml>=4.2.1
Naked>=0.1.31
opuslib>=3.0.1
parse>=1.8.2
parse-type>=0.4.2
pydub>=0.21.0
PyHamcrest>=1.9.0
pyshould>=0.7.1
python-dateutil>=2.7.2
PyYAML>=3.12
requests>=2.18.4
responses>=0.9.0
selenium>=3.11.0
shellescape>=3.4.1
six>=1.11.0
urllib3>=1.22
